hierovision/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
Werkzeug==2.3.7
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
